requirements.txt
# -----------------------------
# FastAPI + Server
# -----------------------------
fastapi>=0.110.0
uvicorn>=0.27.0

# -----------------------------
# Background jobs
# -----------------------------
# celery>=5.3.6
# redis>=5.0.1

# -----------------------------
# PDF processing
# -----------------------------
pypdf>=4.0.0
pdf2image>=1.17.0
pytesseract>=0.3.10
pillow>=10.2.0

# -----------------------------
# Web scraping (STATIC + JS)
# -----------------------------
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
playwright>=1.42.0

# -----------------------------
# LangChain + OpenAI
# -----------------------------
langchain-openai>=0.1.6
langchain-core>=0.2.0
langchain-text-splitters>=0.1.4
tiktoken>=0.6.0
httpx>=0.27.0
openai>=1.10.0

# -----------------------------
# Vector DB
# -----------------------------
pinecone>=3.0.0

# -----------------------------
# Firestore
# -----------------------------
# google-cloud-firestore>=2.14.0
# google-auth>=2.28.0

# -----------------------------
# Utils
# -----------------------------
python-dotenv>=1.0.1
python-multipart>=0.0.9
tenacity>=8.2.3